+++ tests/tests/edgecase.rs
@@ -208,3 +208,62 @@ mod benches {
         )
     }
 }
+
+mod unicode_whitespace {
+    use super::*;
+
+    #[derive(Logos, Debug, Clone, Copy, PartialEq)]
+    #[logos(trivia = r"\p{Whitespace}")]
+    enum Token {
+        #[error]
+        Error,
+
+        #[end]
+        End,
+
+        #[regex = "[0-9]+"]
+        Number,
+    }
+
+    #[test]
+    fn abcdef_trivia() {
+        assert_lex(
+            "   12345\u{2029}67890\t  x ",
+            &[
+                (Token::Number, "12345", 3..8),
+                (Token::Number, "67890", 11..16),
+                (Token::Error, "x", 19..20),
+            ],
+        );
+    }
+}
+
+mod trivia {
+    use super::*;
+
+    #[derive(Logos, Debug, Clone, Copy, PartialEq)]
+    #[logos(trivia = "[a-f]")]
+    enum Token {
+        #[error]
+        Error,
+
+        #[end]
+        End,
+
+        #[regex = "[0-9]+"]
+        Number,
+    }
+
+    #[test]
+    fn abcdef_trivia() {
+        assert_lex(
+            "abc12345def67890 afx",
+            &[
+                (Token::Number, "12345", 3..8),
+                (Token::Number, "67890", 11..16),
+                (Token::Error, " ", 16..17),
+                (Token::Error, "x", 19..20),
+            ],
+        );
+    }
+}
