+++ tests/tests/properties.rs
@@ -1,4 +1,5 @@
 use logos_derive::Logos;
+use tests::assert_lex;
 
 #[derive(Logos, Debug, Clone, Copy, PartialEq)]
 enum Token {
@@ -18,31 +19,26 @@ enum Token {
     Cyrillic,
 }
 
-mod properties {
-    use super::*;
-    use tests::assert_lex;
+#[test]
+fn greek() {
+    assert_lex(
+        "λόγος can do unicode",
+        &[
+            (Token::Greek, "λόγος", 0..10),
+            (Token::Ascii, "can", 11..14),
+            (Token::Ascii, "do", 15..17),
+            (Token::Ascii, "unicode", 18..25),
+        ],
+    )
+}
 
-    #[test]
-    fn greek() {
-        assert_lex(
-            "λόγος can do unicode",
-            &[
-                (Token::Greek, "λόγος", 0..10),
-                (Token::Ascii, "can", 11..14),
-                (Token::Ascii, "do", 15..17),
-                (Token::Ascii, "unicode", 18..25),
-            ],
-        )
-    }
-
-    #[test]
-    fn cyrillic() {
-        assert_lex(
-            "До свидания",
-            &[
-                (Token::Cyrillic, "До", 0..4),
-                (Token::Cyrillic, "свидания", 5..21),
-            ],
-        )
-    }
+#[test]
+fn cyrillic() {
+    assert_lex(
+        "До свидания",
+        &[
+            (Token::Cyrillic, "До", 0..4),
+            (Token::Cyrillic, "свидания", 5..21),
+        ],
+    )
 }
